refactor(datastorage): split workflow_handlers.go by concern (Wave 3)
Split pkg/datastorage/server/workflow_handlers.go (1,687 lines) into
workflow_create_handlers.go (create + validation), workflow_duplicate_handlers.go
(duplicate detection, content integrity, re-enablement), workflow_query_handlers.go
(list/get), workflow_update_lifecycle_handlers.go (update + enable/deprecate/disable
lifecycle), and workflow_discovery_handlers.go (3-step discovery protocol).

Also includes the handleCreateAuditEventsBatch decomposition (cyclomatic 24+ ->
extracted request-parsing/validation/persistence steps), inadvertently omitted
from the earlier decomposition commit.

Pure code motion / behavior-preserving refactor, no exported-API change. Fixed
a missing net/http import in the newly created workflow_duplicate_handlers.go
(http.StatusOK/StatusCreated/StatusConflict references). See audit report
GO-ANTIPATTERN-AUDIT-2026-07-01 §7k-2/§7k-3.
