+++ pkg/datastorage/server/audit_events_batch_handler.go
@@ -84,32 +84,50 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
 	defer cancel()
 
-	// 1. Parse request body as JSON array using OpenAPI type (type-safe)
+	requests, ok := s.decodeAuditEventsBatchRequest(w, r)
+	if !ok {
+		return
+	}
+
+	authenticatedActorID := r.Header.Get("X-Auth-Request-User")
+	auditEvents, repositoryEvents, ok := s.validateAndConvertBatch(w, r, requests, authenticatedActorID)
+	if !ok {
+		return
+	}
+
+	if !s.resolveBatchParentDates(ctx, w, r, requests, auditEvents, repositoryEvents) {
+		return
+	}
+
+	s.warnOutOfOrderTimestamps(auditEvents)
+	s.persistAuditEventsBatch(ctx, w, r, auditEvents, repositoryEvents)
+}
+
+// decodeAuditEventsBatchRequest parses and size-validates the batch request
+// body. Returns ok=false after writing the appropriate RFC 7807 error
+// response if the body is not a valid, non-empty, within-limits JSON array.
+func (s *Server) decodeAuditEventsBatchRequest(w http.ResponseWriter, r *http.Request) ([]dsclient.AuditEventRequest, bool) {
 	var requests []dsclient.AuditEventRequest
 	if err := json.NewDecoder(r.Body).Decode(&requests); err != nil {
 		if dsmiddleware.IsMaxBytesError(err) {
 			dsmiddleware.WriteMaxBytesExceeded(w, s.logger)
-			return
+			return nil, false
 		}
-		// Check if error is due to non-array payload
-		errMsg := err.Error()
-		if strings.Contains(errMsg, "cannot unmarshal object") {
+		if strings.Contains(err.Error(), "cannot unmarshal object") {
 			s.logger.Info("Batch endpoint received single object instead of array", "error", err)
 			response.WriteRFC7807Error(w, http.StatusBadRequest, "invalid_request", "Invalid Request", "request body must be a JSON array, not a single object", s.logger)
-			return
+			return nil, false
 		}
-
 		s.logger.Info("Invalid JSON array in request body", "error", err)
 		response.WriteRFC7807Error(w, http.StatusBadRequest, "invalid_request", "Invalid Request",
 			"The request body must be a valid JSON array of audit event objects", s.logger)
-		return
+		return nil, false
 	}
 
-	// 2. Validate batch is not empty
 	if len(requests) == 0 {
 		s.logger.Info("Batch endpoint received empty array")
 		response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Validation Error", "batch cannot be empty", s.logger)
-		return
+		return nil, false
 	}
 
 	// Issue #667 / BR-STORAGE-043: Enforce maximum batch size to prevent lock amplification
@@ -123,29 +141,36 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 			Detail:   fmt.Sprintf("batch size %d exceeds maximum allowed batch size of %d", len(requests), s.maxBatchSize),
 			Instance: r.URL.Path,
 		}, s)
-		return
+		return nil, false
 	}
 
 	s.logger.V(1).Info("Parsing batch of audit events", "count", len(requests))
+	return requests, true
+}
 
-	authenticatedActorID := r.Header.Get("X-Auth-Request-User")
+// batchParentRef pairs a request's index in the batch with its declared
+// parent_event_id, for the batched FK existence check in
+// resolveBatchParentDates.
+type batchParentRef struct {
+	index    int
+	parentID uuid.UUID
+}
 
-	// 3. Validate and convert ALL events BEFORE persisting any (atomic batch)
+// validateAndConvertBatch validates and converts every request in the batch
+// to internal/repository audit events BEFORE any persistence, preserving the
+// atomic-batch guarantee (an invalid event at any position rejects the whole
+// batch). Returns ok=false after writing the RFC 7807 error response for the
+// first failing event.
+func (s *Server) validateAndConvertBatch(w http.ResponseWriter, r *http.Request, requests []dsclient.AuditEventRequest, authenticatedActorID string) ([]*audit.AuditEvent, []*repository.AuditEvent, bool) {
 	auditEvents := make([]*audit.AuditEvent, 0, len(requests))
 	repositoryEvents := make([]*repository.AuditEvent, 0, len(requests))
-	// PERF-H1: Collect parent IDs for batch FK check instead of N per-row queries.
-	type parentRef struct {
-		index    int
-		parentID uuid.UUID
-	}
-	var parentRefs []parentRef
 
 	for i, req := range requests {
 		if err := helpers.ValidateAuditEventRequest(&req); err != nil {
 			s.logger.Info("Batch validation failed", "index", i, "error", err)
 			response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Validation Error",
 				fmt.Sprintf("event at index %d failed validation", i), s.logger)
-			return
+			return nil, nil, false
 		}
 
 		internalEvent, err := helpers.ConvertAuditEventRequest(req, authenticatedActorID)
@@ -153,18 +178,14 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 			s.logger.Info("Batch conversion failed", "index", i, "error", err)
 			response.WriteRFC7807Error(w, http.StatusBadRequest, "conversion_error", "Conversion Error",
 				fmt.Sprintf("event at index %d could not be converted", i), s.logger)
-			return
+			return nil, nil, false
 		}
 
 		if err := dlq.ValidateEventData(internalEvent.EventData); err != nil {
 			s.logger.Info("Batch EventData validation failed", "index", i, "error", err)
 			response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Validation Error",
 				fmt.Sprintf("event at index %d: event_data exceeds size or nesting depth limits", i), s.logger)
-			return
-		}
-		// Collect parent IDs for batch FK verification below.
-		if req.ParentEventID.IsSet() {
-			parentRefs = append(parentRefs, parentRef{index: i, parentID: req.ParentEventID.Value})
+			return nil, nil, false
 		}
 		auditEvents = append(auditEvents, internalEvent)
 
@@ -173,43 +194,65 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 			s.logger.Info("Batch repository conversion failed - invalid event_data", "index", i, "error", err)
 			response.WriteRFC7807Error(w, http.StatusBadRequest, "invalid_event_data", "Invalid Event Data",
 				fmt.Sprintf("event at index %d has invalid event_data", i), s.logger)
-			return
+			return nil, nil, false
 		}
 		repositoryEvents = append(repositoryEvents, repoEvent)
 	}
 
-	// PERF-H1: Batch FK check -- single query for all parent_event_ids.
-	if len(parentRefs) > 0 {
-		parentIDs := make([]uuid.UUID, 0, len(parentRefs))
-		for _, pr := range parentRefs {
-			parentIDs = append(parentIDs, pr.parentID)
-		}
-		foundParents, fkErr := s.batchLookupParentDates(ctx, parentIDs)
-		if fkErr != nil {
-			s.logger.Error(fkErr, "Batch FK lookup failed")
-			response.WriteRFC7807Error(w, http.StatusInternalServerError,
-				"query-error", "Internal Server Error",
-				"Failed to verify parent events", s.logger)
-			return
-		}
-		for _, pr := range parentRefs {
-			parentDate, ok := foundParents[pr.parentID]
-			if !ok {
-				s.logger.Info("Batch parent event not found", "index", pr.index, "parent_event_id", pr.parentID.String())
-				response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Validation Error",
-					fmt.Sprintf("event at index %d: parent event does not exist", pr.index), s.logger)
-				return
-			}
-			// DF-M2: Propagate resolved parent date into the internal event
-			// so ConvertToRepositoryAuditEvent carries it to the DB layer.
-			auditEvents[pr.index].ParentEventDate = &parentDate
-			repositoryEvents[pr.index].ParentEventDate = &parentDate
+	return auditEvents, repositoryEvents, true
+}
+
+// resolveBatchParentDates performs the PERF-H1 batched FK existence check:
+// a single query resolves event_date for every declared parent_event_id in
+// the batch, instead of one query per row. Returns ok=false after writing
+// the RFC 7807 error response if the lookup query fails or any referenced
+// parent event does not exist.
+func (s *Server) resolveBatchParentDates(ctx context.Context, w http.ResponseWriter, r *http.Request, requests []dsclient.AuditEventRequest, auditEvents []*audit.AuditEvent, repositoryEvents []*repository.AuditEvent) bool {
+	var parentRefs []batchParentRef
+	for i, req := range requests {
+		if req.ParentEventID.IsSet() {
+			parentRefs = append(parentRefs, batchParentRef{index: i, parentID: req.ParentEventID.Value})
 		}
 	}
+	if len(parentRefs) == 0 {
+		return true
+	}
 
-	s.warnOutOfOrderTimestamps(auditEvents)
+	parentIDs := make([]uuid.UUID, 0, len(parentRefs))
+	for _, pr := range parentRefs {
+		parentIDs = append(parentIDs, pr.parentID)
+	}
 
-	// 4. Persist batch atomically (transaction)
+	foundParents, fkErr := s.batchLookupParentDates(ctx, parentIDs)
+	if fkErr != nil {
+		s.logger.Error(fkErr, "Batch FK lookup failed")
+		response.WriteRFC7807Error(w, http.StatusInternalServerError,
+			"query-error", "Internal Server Error",
+			"Failed to verify parent events", s.logger)
+		return false
+	}
+
+	for _, pr := range parentRefs {
+		parentDate, ok := foundParents[pr.parentID]
+		if !ok {
+			s.logger.Info("Batch parent event not found", "index", pr.index, "parent_event_id", pr.parentID.String())
+			response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Validation Error",
+				fmt.Sprintf("event at index %d: parent event does not exist", pr.index), s.logger)
+			return false
+		}
+		// DF-M2: Propagate resolved parent date into the internal event
+		// so ConvertToRepositoryAuditEvent carries it to the DB layer.
+		auditEvents[pr.index].ParentEventDate = &parentDate
+		repositoryEvents[pr.index].ParentEventDate = &parentDate
+	}
+	return true
+}
+
+// persistAuditEventsBatch writes the batch atomically via CreateBatch. On
+// success, it responds 201 with the created event_ids. On DB failure, it
+// falls back to per-item DLQ enqueue (DD-009) and responds 202 if every item
+// was queued, or 500 if any item was lost.
+func (s *Server) persistAuditEventsBatch(ctx context.Context, w http.ResponseWriter, r *http.Request, auditEvents []*audit.AuditEvent, repositoryEvents []*repository.AuditEvent) {
 	s.logger.V(1).Info("Writing batch to database", "count", len(repositoryEvents))
 
 	start := time.Now()
@@ -220,53 +263,10 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 	s.metrics.WriteDuration.WithLabelValues("audit_events_batch").Observe(duration)
 
 	if err != nil {
-		s.logger.Error(err, "Batch database write failed, attempting per-item DLQ fallback",
-			"count", len(auditEvents),
-			"duration_seconds", duration)
-
-		// DD-009: DLQ fallback on database errors (per-item, mirroring single-event path)
-		dlqCtx, dlqCancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer dlqCancel()
-
-		var dlqSuccessCount, dlqFailCount int
-		for i, auditEvent := range auditEvents {
-			if dlqErr := s.dlqClient.EnqueueAuditEvent(dlqCtx, auditEvent, err); dlqErr != nil {
-				s.logger.Error(dlqErr, "DLQ fallback failed for batch item — data loss risk",
-					"index", i,
-					"event_type", auditEvent.EventType,
-					"correlation_id", auditEvent.CorrelationID)
-				dlqFailCount++
-			} else {
-				dlqSuccessCount++
-			}
-		}
-
-		if dlqFailCount > 0 {
-			s.logger.Error(err, "Batch DLQ fallback partially failed",
-				"dlq_success", dlqSuccessCount,
-				"dlq_failed", dlqFailCount,
-				"total", len(auditEvents))
-			writeValidationRFC7807Error(w, &validation.RFC7807Problem{
-				Type:     "https://kubernaut.ai/problems/database-error",
-				Title:    "Database Error",
-				Status:   http.StatusInternalServerError,
-				Detail:   fmt.Sprintf("Batch write failed; %d of %d events queued to DLQ, %d lost", dlqSuccessCount, len(auditEvents), dlqFailCount),
-				Instance: r.URL.Path,
-			}, s)
-			return
-		}
-
-		s.logger.Info("Batch DLQ fallback succeeded — all events queued",
-			"count", dlqSuccessCount)
-		response.WriteJSON(w, http.StatusAccepted, BatchAuditEventAcceptedResponse{
-			Status:  "accepted",
-			Message: fmt.Sprintf("%d audit events queued for async processing via DLQ", dlqSuccessCount),
-			Count:   dlqSuccessCount,
-		}, s.logger)
+		s.handleBatchWriteFailure(w, r, auditEvents, err, duration)
 		return
 	}
 
-	// 5. Build response with event_ids
 	eventIDs := make([]string, len(createdEvents))
 	for i, event := range createdEvents {
 		eventIDs[i] = event.EventID.String()
@@ -276,7 +276,6 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 		"count", len(eventIDs),
 		"duration_seconds", duration)
 
-	// 7. Return 201 Created
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusCreated)
 	resp := BatchAuditEventCreatedResponse{
@@ -286,6 +285,55 @@ func (s *Server) handleCreateAuditEventsBatch(w http.ResponseWriter, r *http.Req
 	if err := json.NewEncoder(w).Encode(resp); err != nil {
 		s.logger.Error(err, "failed to encode batch response")
 	}
+}
+
+// handleBatchWriteFailure implements the DD-009 per-item DLQ fallback when
+// the atomic batch write to the database fails: every event is individually
+// enqueued to the DLQ for async retry so no event is lost on a transient DB
+// outage. Responds 202 if all events were queued, 500 if any were lost.
+func (s *Server) handleBatchWriteFailure(w http.ResponseWriter, r *http.Request, auditEvents []*audit.AuditEvent, writeErr error, duration float64) {
+	s.logger.Error(writeErr, "Batch database write failed, attempting per-item DLQ fallback",
+		"count", len(auditEvents),
+		"duration_seconds", duration)
+
+	dlqCtx, dlqCancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer dlqCancel()
+
+	var dlqSuccessCount, dlqFailCount int
+	for i, auditEvent := range auditEvents {
+		if dlqErr := s.dlqClient.EnqueueAuditEvent(dlqCtx, auditEvent, writeErr); dlqErr != nil {
+			s.logger.Error(dlqErr, "DLQ fallback failed for batch item — data loss risk",
+				"index", i,
+				"event_type", auditEvent.EventType,
+				"correlation_id", auditEvent.CorrelationID)
+			dlqFailCount++
+		} else {
+			dlqSuccessCount++
+		}
+	}
+
+	if dlqFailCount > 0 {
+		s.logger.Error(writeErr, "Batch DLQ fallback partially failed",
+			"dlq_success", dlqSuccessCount,
+			"dlq_failed", dlqFailCount,
+			"total", len(auditEvents))
+		writeValidationRFC7807Error(w, &validation.RFC7807Problem{
+			Type:     "https://kubernaut.ai/problems/database-error",
+			Title:    "Database Error",
+			Status:   http.StatusInternalServerError,
+			Detail:   fmt.Sprintf("Batch write failed; %d of %d events queued to DLQ, %d lost", dlqSuccessCount, len(auditEvents), dlqFailCount),
+			Instance: r.URL.Path,
+		}, s)
+		return
+	}
+
+	s.logger.Info("Batch DLQ fallback succeeded — all events queued",
+		"count", dlqSuccessCount)
+	response.WriteJSON(w, http.StatusAccepted, BatchAuditEventAcceptedResponse{
+		Status:  "accepted",
+		Message: fmt.Sprintf("%d audit events queued for async processing via DLQ", dlqSuccessCount),
+		Count:   dlqSuccessCount,
+	}, s.logger)
 }
 
 // warnOutOfOrderTimestamps logs when events in the same batch share a correlation_id
