--- pkg/datastorage/server/workflow_handlers.go
@@ -1,1714 +1,0 @@
-/*
-Copyright 2025 Jordi Gil.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package server
-
-import (
-	"context"
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/http"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/go-chi/chi/v5"
-	"github.com/go-logr/logr"
-	"github.com/jackc/pgx/v5/pgconn"
-	dsaudit "github.com/jordigilh/kubernaut/pkg/datastorage/audit"
-	dsmetrics "github.com/jordigilh/kubernaut/pkg/datastorage/metrics"
-	"github.com/jordigilh/kubernaut/pkg/datastorage/models"
-	api "github.com/jordigilh/kubernaut/pkg/datastorage/ogen-client"
-	"github.com/jordigilh/kubernaut/pkg/datastorage/schema"
-	dsmiddleware "github.com/jordigilh/kubernaut/pkg/datastorage/server/middleware"
-	"github.com/jordigilh/kubernaut/pkg/datastorage/server/response"
-	deterministicuuid "github.com/jordigilh/kubernaut/pkg/datastorage/uuid"
-)
-
-// ========================================
-// WORKFLOW CATALOG HANDLERS
-// ========================================
-// Business Requirements:
-// - BR-STORAGE-014: Workflow catalog management
-// - BR-STORAGE-039: Workflow Catalog Retrieval API
-//
-// API Endpoints:
-// - POST /api/v1/workflows - Create a new workflow
-// - GET /api/v1/workflows - List workflows with filters
-// - GET /api/v1/workflows/{workflowID} - Get workflow by ID (with optional security gate)
-// - PATCH /api/v1/workflows/{workflowID} - Update mutable fields
-// - PATCH /api/v1/workflows/{workflowID}/disable - Soft-delete workflow
-//
-// Three-Step Discovery Protocol (DD-HAPI-017):
-// - GET /api/v1/workflows/actions - Step 1: List available action types
-// - GET /api/v1/workflows/actions/{action_type} - Step 2: List workflows by action type
-// - GET /api/v1/workflows/{workflowID} - Step 3: Get workflow with security gate
-
-// HandleCreateWorkflow handles POST /api/v1/workflows
-// ADR-058: Inline workflow schema registration (CRD-based)
-// BR-WORKFLOW-006: Accept content (raw YAML), parse CRD envelope, validate, populate catalog
-func (h *Handler) HandleCreateWorkflow(w http.ResponseWriter, r *http.Request) {
-	// Step 1: Parse request body — expect {"content": "...", "source": "...", "registeredBy": "..."}
-	// Cap request body at 2 MiB to prevent memory exhaustion from oversized payloads.
-	const maxBodySize = 2 << 20 // 2 MiB
-	r.Body = http.MaxBytesReader(w, r.Body, maxBodySize)
-
-	var req struct {
-		Content      string `json:"content"`
-		Source       string `json:"source"`
-		RegisteredBy string `json:"registeredBy"`
-		SchemaImage  string `json:"schemaImage"` // legacy field — reject if present
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		if dsmiddleware.IsMaxBytesError(err) {
-			dsmiddleware.WriteMaxBytesExceeded(w, h.logger)
-			return
-		}
-		h.logger.Error(err, "Failed to decode workflow create request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"request body is not valid JSON", h.logger)
-		return
-	}
-
-	// Step 2a: Reject legacy OCI format — guide to new inline format
-	if req.SchemaImage != "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "legacy-format", "Legacy Format Rejected",
-			"The 'schemaImage' field is no longer supported. Use 'content' with the raw YAML of the RemediationWorkflow CRD instead (ADR-058).",
-			h.logger)
-		return
-	}
-
-	// Step 2b: Validate content is present
-	if req.Content == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"content is required", h.logger)
-		return
-	}
-
-	// Step 3: Parse and validate the inline schema
-	schemaParser := schema.NewParser()
-	parsedSchema, err := schemaParser.ParseAndValidate(req.Content)
-	if err != nil {
-		h.logger.Error(err, "Inline schema validation failed")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "validation-error", "Schema Validation Error",
-			"workflow schema validation failed; check the 'content' field for YAML/structural errors", h.logger)
-		return
-	}
-
-	// Step 4: Build RemediationWorkflow from parsed schema (inline — no OCI image)
-	workflow, err := h.buildWorkflowFromInlineSchema(schemaParser, parsedSchema, req.Content)
-	if err != nil {
-		h.logger.Error(err, "Failed to build workflow from inline schema")
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to process inline schema", h.logger)
-		return
-	}
-
-	// Steps 5a–5c: Validate external checks in parallel (Issue #1070)
-	// Typed-result-slot pattern: each check writes to its own error slot,
-	// then we check slots in priority order so callers always see the same
-	// RFC 7807 error type regardless of goroutine completion order.
-	if err := h.validateExternalChecks(r.Context(), schemaParser, parsedSchema, workflow); err != nil {
-		err.writeTo(w, h.logger)
-		return
-	}
-
-	// Step 6: Create workflow in repository (with content integrity checking)
-	// BR-WORKFLOW-006: ContentHash-based duplicate detection prevents spec tampering
-	statusCode := http.StatusCreated
-	if h.workflowIntegrityRepo != nil {
-		result, integrityErr := h.handleDuplicateWorkflow(r.Context(), workflow)
-		if integrityErr != nil {
-			var cie *contentIntegrityError
-			if errors.As(integrityErr, &cie) {
-				h.logger.Info("Content integrity violation: same version with different content",
-					"workflow_name", cie.WorkflowName,
-					"version", cie.Version,
-				)
-				response.WriteRFC7807Error(w, http.StatusConflict, "content-integrity-violation",
-					"Content Changed Without Version Bump", cie.Error(), h.logger)
-				return
-			}
-			h.logger.Error(integrityErr, "Content integrity check failed",
-				"workflow_name", workflow.WorkflowName,
-				"version", workflow.Version,
-			)
-			response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-				"Failed to process workflow registration", h.logger)
-			return
-		}
-		if result != nil {
-			workflow = result.workflow
-			statusCode = result.statusCode
-		}
-	} else if h.workflowRepo != nil {
-		if err := h.workflowRepo.Create(r.Context(), workflow); err != nil {
-			var pgErr *pgconn.PgError
-			if errors.As(err, &pgErr) && pgErr.Code == "23505" {
-				reEnabled, reEnableErr := h.tryReEnableWorkflow(r.Context(), workflow)
-				if reEnableErr != nil {
-					h.logger.Error(reEnableErr, "Failed to re-enable workflow",
-						"workflow_name", workflow.WorkflowName,
-						"version", workflow.Version,
-					)
-					detail := fmt.Sprintf("Workflow '%s' version '%s' already exists in active state", workflow.WorkflowName, workflow.Version)
-					response.WriteRFC7807Error(w, http.StatusConflict, "conflict",
-						"Workflow Already Exists", detail, h.logger)
-					return
-				}
-				if reEnabled != nil {
-					workflow = reEnabled
-					statusCode = http.StatusOK
-				}
-			} else {
-				h.logger.Error(err, "Failed to create workflow",
-					"workflow_name", workflow.WorkflowName,
-					"version", workflow.Version,
-				)
-				response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-					"Failed to create workflow", h.logger)
-				return
-			}
-		}
-	}
-
-	// Step 7: Audit workflow creation (synchronous per ADR-032)
-	// DS-H5: Bounded context (5s) prevents indefinite blocking if Data Storage is slow.
-	if h.auditStore != nil {
-		auditEvent, auditErr := dsaudit.NewWorkflowCreatedAuditEvent(workflow)
-		if auditErr != nil {
-			h.logger.Error(auditErr, "Failed to create workflow creation audit event",
-				"workflow_id", workflow.WorkflowID,
-				"workflow_name", workflow.WorkflowName,
-			)
-		} else {
-			auditCtx, auditCancel := context.WithTimeout(r.Context(), 5*time.Second)
-			defer auditCancel()
-			if storeErr := h.auditStore.StoreAudit(auditCtx, auditEvent); storeErr != nil {
-				h.logger.Error(storeErr, "Failed to persist workflow creation audit",
-					"workflow_id", workflow.WorkflowID,
-					"workflow_name", workflow.WorkflowName,
-				)
-			}
-		}
-	}
-
-	// Log success
-	h.logger.Info("Workflow registered successfully",
-		"workflow_id", workflow.WorkflowID,
-		"workflow_name", workflow.WorkflowName,
-		"version", workflow.Version,
-		"source", req.Source,
-		"registered_by", req.RegisteredBy,
-		"status_code", statusCode,
-	)
-
-	// Return workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(statusCode)
-	if err := json.NewEncoder(w).Encode(workflow); err != nil {
-		h.logger.Error(err, "Failed to encode workflow create response")
-	}
-}
-
-// validationError carries all fields needed to produce an RFC 7807 response.
-// Returned by validateExternalChecks so HandleCreateWorkflow can write the error
-// without duplicating RFC 7807 construction logic.
-type validationError struct {
-	status    int
-	errorType string
-	title     string
-	detail    string
-}
-
-// writeTo writes the validationError as an RFC 7807 Problem Details response.
-func (ve *validationError) writeTo(w http.ResponseWriter, logger logr.Logger) {
-	response.WriteRFC7807Error(w, ve.status, ve.errorType, ve.title, ve.detail, logger)
-}
-
-// validateExternalChecks runs Steps 5a–5c (action-type, bundle-exists,
-// dependency validation) in parallel and returns the highest-priority
-// error, preserving the original sequential error contract.
-//
-// A 10-second timeout budget bounds the total wall-clock time for all
-// external calls, preventing a degraded backend from consuming the
-// entire server WriteTimeout.
-//
-// Issue #1070: Typed-result-slot pattern — each goroutine writes to its
-// own slot; after all goroutines complete we check slots in priority order.
-func (h *Handler) validateExternalChecks(
-	ctx context.Context,
-	schemaParser *schema.Parser,
-	parsedSchema *models.WorkflowSchema,
-	workflow *models.RemediationWorkflow,
-) *validationError {
-	const (
-		slotActionType = iota
-		slotBundle
-		slotDependency
-		slotCount
-
-		validationTimeout = 10 * time.Second
-	)
-
-	ctx, cancel := context.WithTimeout(ctx, validationTimeout)
-	defer cancel()
-
-	var (
-		slots [slotCount]*validationError
-		mu    sync.Mutex
-		wg    sync.WaitGroup
-	)
-
-	totalStart := time.Now()
-
-	setSlot := func(idx int, ve *validationError) {
-		mu.Lock()
-		slots[idx] = ve
-		mu.Unlock()
-	}
-
-	// 5a: Validate action_type against taxonomy (GAP-4, DD-WORKFLOW-016)
-	if h.actionTypeValidator != nil {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			start := time.Now()
-			exists, err := h.actionTypeValidator.ActionTypeExists(ctx, workflow.ActionType)
-			if err != nil {
-				dsmetrics.WorkflowValidationDuration.WithLabelValues("action_type", "error").Observe(time.Since(start).Seconds())
-				h.logger.Error(err, "Failed to validate action_type against taxonomy",
-					"action_type", workflow.ActionType,
-				)
-				setSlot(slotActionType, &validationError{
-					status:    http.StatusInternalServerError,
-					errorType: "internal-error",
-					title:     "Internal Server Error",
-					detail:    "Failed to validate action type",
-				})
-				return
-			}
-			if !exists {
-				dsmetrics.WorkflowValidationDuration.WithLabelValues("action_type", "error").Observe(time.Since(start).Seconds())
-				setSlot(slotActionType, &validationError{
-					status:    http.StatusBadRequest,
-					errorType: "validation-error",
-					title:     "Validation Error",
-					detail:    fmt.Sprintf("action_type '%s' is not in the action type taxonomy (DD-WORKFLOW-016)", workflow.ActionType),
-				})
-				return
-			}
-			dsmetrics.WorkflowValidationDuration.WithLabelValues("action_type", "ok").Observe(time.Since(start).Seconds())
-		}()
-	}
-
-	// 5b: Validate execution bundle image exists in the registry (skip for ansible — Git repo)
-	if workflow.ExecutionBundle != nil && *workflow.ExecutionBundle != "" &&
-		workflow.ExecutionEngine != models.ExecutionEngineAnsible && h.schemaExtractor != nil {
-		bundleRef := *workflow.ExecutionBundle
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			start := time.Now()
-			if err := h.schemaExtractor.ValidateBundleExists(ctx, bundleRef); err != nil {
-				dsmetrics.WorkflowValidationDuration.WithLabelValues("bundle_exists", "error").Observe(time.Since(start).Seconds())
-				h.logger.Error(err, "Execution bundle image not found in registry",
-					"execution_bundle", bundleRef,
-				)
-				setSlot(slotBundle, &validationError{
-					status:    http.StatusBadRequest,
-					errorType: "bundle-not-found",
-					title:     "Execution Bundle Not Found",
-					detail:    "execution.bundle image could not be resolved; verify the image reference is correct",
-				})
-				return
-			}
-			dsmetrics.WorkflowValidationDuration.WithLabelValues("bundle_exists", "ok").Observe(time.Since(start).Seconds())
-		}()
-	}
-
-	// 5c: Validate schema-declared dependencies exist in execution namespace (DD-WE-006)
-	if h.dependencyValidator != nil && parsedSchema.Dependencies != nil {
-		deps := schemaParser.ExtractDependencies(parsedSchema)
-		if deps != nil {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				start := time.Now()
-				if err := h.dependencyValidator.ValidateDependencies(ctx, h.executionNamespace, deps); err != nil {
-					dsmetrics.WorkflowValidationDuration.WithLabelValues("dependency", "error").Observe(time.Since(start).Seconds())
-					h.logger.Error(err, "Dependency validation failed",
-						"execution_namespace", h.executionNamespace,
-					)
-					setSlot(slotDependency, &validationError{
-						status:    http.StatusBadRequest,
-						errorType: "dependency-validation-error",
-						title:     "Dependency Validation Error",
-					detail: fmt.Sprintf("Schema-declared dependency not satisfied in namespace %q; "+
-						"ensure all dependencies are provisioned before registering the workflow (DD-WE-006)", h.executionNamespace),
-					})
-					return
-				}
-				dsmetrics.WorkflowValidationDuration.WithLabelValues("dependency", "ok").Observe(time.Since(start).Seconds())
-			}()
-		}
-	}
-
-	wg.Wait()
-
-	// Return the highest-priority error (lowest slot index).
-	for _, slot := range slots {
-		if slot != nil {
-			dsmetrics.WorkflowValidationDuration.WithLabelValues("total", "error").Observe(time.Since(totalStart).Seconds())
-			return slot
-		}
-	}
-	dsmetrics.WorkflowValidationDuration.WithLabelValues("total", "ok").Observe(time.Since(totalStart).Seconds())
-	return nil
-}
-
-// contentIntegrityError is returned when an active workflow with the same
-// (name, version) already exists but has a different content hash. The caller
-// must bump the version to register new content. This enforces version-locked
-// content immutability per Issue #773.
-type contentIntegrityError struct {
-	WorkflowName string
-	Version      string
-	OldHash      string
-	NewHash      string
-}
-
-func (e *contentIntegrityError) Error() string {
-	oldPrefix, newPrefix := e.OldHash, e.NewHash
-	if len(oldPrefix) > 12 {
-		oldPrefix = oldPrefix[:12]
-	}
-	if len(newPrefix) > 12 {
-		newPrefix = newPrefix[:12]
-	}
-	return fmt.Sprintf(
-		"active workflow %q version %q already has different content (hash %s→%s); bump the version to register new content",
-		e.WorkflowName, e.Version, oldPrefix, newPrefix,
-	)
-}
-
-// duplicateResult holds the outcome of content integrity checking.
-type duplicateResult struct {
-	workflow   *models.RemediationWorkflow
-	statusCode int
-}
-
-// handleDuplicateWorkflow implements ContentHash-based duplicate detection.
-// BR-WORKFLOW-006: Determines the correct action when a workflow with the same
-// name+version already exists: idempotent return, supersede, re-enable, or create new.
-//
-// Decision tree:
-//   active  + same hash    → 200 (idempotent, no DB writes)
-//   active  + diff hash    → 409 contentIntegrityError (must bump version — Issue #773)
-//   disabled + same hash   → re-enable → 200
-//   disabled + diff hash   → create new → 201
-//   cross-version (any)    → supersede old → create new → 201
-//   none                   → create new → 201
-//
-// Concurrency safety: All Create calls handle PostgreSQL 23505 (unique constraint
-// violation on uq_workflow_name_version_active) by retrying the active-lookup.
-// This handles the race where two concurrent requests both pass the initial lookup,
-// one wins the INSERT, and the other must discover the winner's committed row.
-func (h *Handler) handleDuplicateWorkflow(ctx context.Context, workflow *models.RemediationWorkflow) (*duplicateResult, error) {
-	repo := h.workflowIntegrityRepo
-	incomingHash := computeContentHash(workflow.Content)
-	workflow.ContentHash = incomingHash
-	// Issue #548: Derive workflow_id deterministically from content hash so that
-	// re-registering the same CRD content after a PVC wipe recovers the original ID.
-	workflow.WorkflowID = deterministicuuid.DeterministicUUID(incomingHash)
-
-	active, err := repo.GetActiveByNameAndVersion(ctx, workflow.WorkflowName, workflow.Version)
-	if err != nil {
-		return nil, fmt.Errorf("lookup active workflow: %w", err)
-	}
-
-	if active != nil {
-		if active.ContentHash == incomingHash {
-			h.logger.Info("Idempotent workflow re-apply (same content hash)",
-				"workflow_id", active.WorkflowID,
-				"workflow_name", active.WorkflowName,
-				"content_hash", incomingHash,
-			)
-			return &duplicateResult{workflow: active, statusCode: http.StatusOK}, nil
-		}
-
-		return nil, &contentIntegrityError{
-			WorkflowName: active.WorkflowName,
-			Version:      active.Version,
-			OldHash:      active.ContentHash,
-			NewHash:      incomingHash,
-		}
-	}
-
-	disabled, err := repo.GetLatestDisabledByNameAndVersion(ctx, workflow.WorkflowName, workflow.Version)
-	if err != nil {
-		return nil, fmt.Errorf("lookup disabled workflow: %w", err)
-	}
-
-	if disabled != nil {
-		if disabled.ContentHash == incomingHash {
-			if err := repo.UpdateStatus(ctx, disabled.WorkflowID, disabled.Version, "Active", "re-enabled via CRD re-creation", ""); err != nil {
-				return nil, fmt.Errorf("re-enable workflow %s: %w", disabled.WorkflowID, err)
-			}
-			disabled.Status = "Active"
-			disabled.DisabledAt = nil
-			disabled.DisabledBy = nil
-			disabled.DisabledReason = nil
-
-			h.logger.Info("Disabled workflow re-enabled (same content hash)",
-				"workflow_id", disabled.WorkflowID,
-				"workflow_name", disabled.WorkflowName,
-				"content_hash", incomingHash,
-			)
-			return &duplicateResult{workflow: disabled, statusCode: http.StatusOK}, nil
-		}
-
-		if err := repo.Create(ctx, workflow); err != nil {
-			if result, handled := h.retryOnUniqueViolation(ctx, err, workflow, incomingHash); handled {
-				return result, nil
-			}
-			return nil, fmt.Errorf("create new workflow (disabled had different content): %w", err)
-		}
-
-		h.logger.Info("New workflow created (disabled had different content hash)",
-			"workflow_id", workflow.WorkflowID,
-			"workflow_name", workflow.WorkflowName,
-			"disabled_id", disabled.WorkflowID,
-		)
-		return &duplicateResult{workflow: workflow, statusCode: http.StatusCreated}, nil
-	}
-
-	// Issue #371: Cross-version check — look for any active entry with the same
-	// workflow_name (regardless of version). If found, the incoming request is a
-	// version upgrade and the old entry must be superseded to enforce single-active
-	// per (workflow_name, action_type).
-	activeAnyVersion, err := repo.GetActiveByWorkflowName(ctx, workflow.WorkflowName)
-	if err != nil {
-		return nil, fmt.Errorf("lookup active workflow by name: %w", err)
-	}
-
-	if activeAnyVersion != nil {
-		reason := fmt.Sprintf("superseded: new version %s registered (was %s)", workflow.Version, activeAnyVersion.Version)
-		if err := repo.SupersedeAndCreate(ctx, activeAnyVersion.WorkflowID, activeAnyVersion.Version, reason, workflow); err != nil {
-			if result, handled := h.retryOnUniqueViolation(ctx, err, workflow, incomingHash); handled {
-				return result, nil
-			}
-			return nil, fmt.Errorf("supersede and create after cross-version update: %w", err)
-		}
-
-		h.logger.Info("Workflow superseded due to cross-version update (Issue #371)",
-			"old_workflow_id", activeAnyVersion.WorkflowID,
-			"old_version", activeAnyVersion.Version,
-			"workflow_name", activeAnyVersion.WorkflowName,
-			"new_version", workflow.Version,
-		)
-		return &duplicateResult{workflow: workflow, statusCode: http.StatusCreated}, nil
-	}
-
-	if err := repo.Create(ctx, workflow); err != nil {
-		if result, handled := h.retryOnUniqueViolation(ctx, err, workflow, incomingHash); handled {
-			return result, nil
-		}
-		return nil, fmt.Errorf("create new workflow: %w", err)
-	}
-	return &duplicateResult{workflow: workflow, statusCode: http.StatusCreated}, nil
-}
-
-// retryOnUniqueViolation handles PostgreSQL 23505 errors from concurrent Create
-// calls by re-querying for the active workflow that won the INSERT race.
-// Returns (result, true) if the 23505 was successfully handled, or (nil, false)
-// if the error is not a 23505 and should be propagated by the caller.
-func (h *Handler) retryOnUniqueViolation(ctx context.Context, createErr error, workflow *models.RemediationWorkflow, incomingHash string) (*duplicateResult, bool) {
-	var pgErr *pgconn.PgError
-	if !errors.As(createErr, &pgErr) || pgErr.Code != "23505" {
-		return nil, false
-	}
-
-	h.logger.Info("Concurrent INSERT race detected (23505), retrying active-lookup",
-		"workflow_name", workflow.WorkflowName,
-		"version", workflow.Version,
-		"constraint", pgErr.ConstraintName,
-	)
-
-	active, err := h.workflowIntegrityRepo.GetActiveByNameAndVersion(ctx, workflow.WorkflowName, workflow.Version)
-	if err != nil {
-		h.logger.Error(err, "Retry lookup failed after 23505",
-			"workflow_name", workflow.WorkflowName, "version", workflow.Version)
-		return nil, false
-	}
-	if active == nil {
-		h.logger.Error(nil, "No active workflow found on retry after 23505 — constraint violation from unexpected source",
-			"workflow_name", workflow.WorkflowName, "version", workflow.Version)
-		return nil, false
-	}
-
-	if active.ContentHash == incomingHash {
-		h.logger.Info("Concurrent race resolved: idempotent (same content hash)",
-			"workflow_id", active.WorkflowID,
-			"workflow_name", active.WorkflowName,
-			"content_hash", incomingHash,
-		)
-		return &duplicateResult{workflow: active, statusCode: http.StatusOK}, true
-	}
-
-	h.logger.Info("Concurrent race resolved: different content already active",
-		"active_workflow_id", active.WorkflowID,
-		"active_hash", active.ContentHash,
-		"incoming_hash", incomingHash,
-	)
-	return &duplicateResult{workflow: active, statusCode: http.StatusConflict}, true
-}
-
-// tryReEnableWorkflow attempts to re-enable a previously disabled workflow on duplicate insert.
-// Returns the re-enabled workflow on success, or an error if the workflow is active.
-func (h *Handler) tryReEnableWorkflow(ctx context.Context, workflow *models.RemediationWorkflow) (*models.RemediationWorkflow, error) {
-	repo := h.getWorkflowLifecycleRepo()
-	if repo == nil {
-		return nil, fmt.Errorf("workflow repository not configured for re-enable")
-	}
-
-	existing, err := h.workflowRepo.GetByNameAndVersion(ctx, workflow.WorkflowName, workflow.Version)
-	if err != nil {
-		return nil, fmt.Errorf("lookup existing workflow: %w", err)
-	}
-	if existing == nil {
-		return nil, fmt.Errorf("workflow not found after conflict")
-	}
-
-	if existing.Status != "Disabled" {
-		return nil, fmt.Errorf("workflow is %s, not disabled", existing.Status)
-	}
-
-	reason := "re-enabled via CRD re-creation"
-	if err := repo.UpdateStatus(ctx, existing.WorkflowID, existing.Version, "Active", reason, ""); err != nil {
-		return nil, fmt.Errorf("update status to active: %w", err)
-	}
-
-	existing.Status = "Active"
-	existing.DisabledAt = nil
-	existing.DisabledBy = nil
-	existing.DisabledReason = nil
-	return existing, nil
-}
-
-// buildWorkflowFromInlineSchema populates a RemediationWorkflow from inline CRD YAML.
-// ADR-058: SchemaImage and SchemaDigest are nil for inline registration.
-func (h *Handler) buildWorkflowFromInlineSchema(
-	schemaParser *schema.Parser,
-	parsedSchema *models.WorkflowSchema,
-	rawContent string,
-) (*models.RemediationWorkflow, error) {
-	return h.buildWorkflowCommon(schemaParser, parsedSchema, rawContent)
-}
-
-// buildWorkflowCommon is the shared workflow builder for both OCI and inline registration.
-func (h *Handler) buildWorkflowCommon(
-	schemaParser *schema.Parser,
-	parsedSchema *models.WorkflowSchema,
-	rawContent string,
-) (*models.RemediationWorkflow, error) {
-	extractedParams, err := schemaParser.ExtractParameters(parsedSchema)
-	if err != nil {
-		return nil, fmt.Errorf("extract parameters: %w", err)
-	}
-	wrappedParams := map[string]interface{}{
-		"schema": map[string]json.RawMessage{
-			"parameters": extractedParams,
-		},
-	}
-	wrappedJSON, err := json.Marshal(wrappedParams)
-	if err != nil {
-		return nil, fmt.Errorf("marshal parameters: %w", err)
-	}
-	rawParams := json.RawMessage(wrappedJSON)
-
-	labelsJSON, err := schemaParser.ExtractLabels(parsedSchema)
-	if err != nil {
-		return nil, fmt.Errorf("extract labels: %w", err)
-	}
-
-	desc := models.StructuredDescription{
-		What:          parsedSchema.Description.What,
-		WhenToUse:     parsedSchema.Description.WhenToUse,
-		WhenNotToUse:  parsedSchema.Description.WhenNotToUse,
-		Preconditions: parsedSchema.Description.Preconditions,
-	}
-
-	execEngine := models.ExecutionEngine(schemaParser.ExtractExecutionEngine(parsedSchema))
-
-	workflow := &models.RemediationWorkflow{
-		WorkflowName:    parsedSchema.WorkflowName,
-		Version:         parsedSchema.Version,
-		SchemaVersion:   parsedSchema.SchemaVersion,
-		Name:            parsedSchema.WorkflowName,
-		Description:     desc,
-		Content:         rawContent,
-		Parameters:      &rawParams,
-		ExecutionEngine: execEngine,
-		ActionType:      parsedSchema.ActionType,
-		Status:          "Active",
-		IsLatestVersion: true,
-	}
-
-	if bundle := schemaParser.ExtractExecutionBundle(parsedSchema); bundle != nil {
-		workflow.ExecutionBundle = bundle
-		if _, digest, err := schema.ParseBundleDigest(*bundle); err == nil {
-			workflow.ExecutionBundleDigest = &digest
-		}
-	}
-
-	workflow.EngineConfig = schemaParser.ExtractEngineConfig(parsedSchema)
-	workflow.ServiceAccountName = schemaParser.ExtractServiceAccountName(parsedSchema)
-
-	if err := json.Unmarshal(labelsJSON, &workflow.Labels); err != nil {
-		return nil, fmt.Errorf("unmarshal labels: %w", err)
-	}
-
-	detectedLabels, err := schemaParser.ExtractDetectedLabels(parsedSchema)
-	if err != nil {
-		return nil, fmt.Errorf("extract detected labels: %w", err)
-	}
-	workflow.DetectedLabels = *detectedLabels
-
-	workflow.CustomLabels = schemaParser.ExtractCustomLabels(parsedSchema)
-	workflow.ContentHash = computeContentHash(rawContent)
-
-	return workflow, nil
-}
-
-// HandleListWorkflows handles GET /api/v1/workflows
-// BR-STORAGE-014: Workflow catalog management
-func (h *Handler) HandleListWorkflows(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters
-	filters := &models.WorkflowSearchFilters{}
-
-	// Status filter
-	if status := r.URL.Query().Get("status"); status != "" {
-		filters.Status = []string{status}
-	}
-
-	// DD-WORKFLOW-001 v2.4: Multi-environment support (accepts multiple values)
-	// Parse environment filter from query parameters
-	// DD-WORKFLOW-001 v2.5: Single environment value (workflows store arrays, searches use single value)
-	if env := r.URL.Query().Get("environment"); env != "" {
-		filters.Environment = env
-	}
-
-	// Priority filter (mandatory)
-	if priority := r.URL.Query().Get("priority"); priority != "" {
-		filters.Priority = priority
-	}
-
-	// Component filter (mandatory)
-	if component := r.URL.Query().Get("component"); component != "" {
-		filters.Component = component
-	}
-
-	// Workflow name filter (exact match for metadata lookup)
-	// Authority: DD-API-001 (OpenAPI client mandatory - added in Jan 2026)
-	// Used for test idempotency and workflow lookup by human-readable name
-	if workflowName := r.URL.Query().Get("workflow_name"); workflowName != "" {
-		filters.WorkflowName = workflowName
-	}
-
-	// DD-WORKFLOW-001 v1.5: Custom labels (subdomain-based)
-	// Format: custom_labels[subdomain]=value1,value2
-	// Example: custom_labels[constraint]=cost-constrained,stateful-safe
-	for key, values := range r.URL.Query() {
-		if strings.HasPrefix(key, "custom_labels[") && strings.HasSuffix(key, "]") {
-			subdomain := strings.TrimSuffix(strings.TrimPrefix(key, "custom_labels["), "]")
-			if subdomain != "" && len(values) > 0 {
-				if filters.CustomLabels == nil {
-					filters.CustomLabels = make(map[string][]string)
-				}
-				// Split comma-separated values
-				for _, v := range values {
-					filters.CustomLabels[subdomain] = append(filters.CustomLabels[subdomain], strings.Split(v, ",")...)
-				}
-			}
-		}
-	}
-
-	// Pagination
-	limit, err := strconv.Atoi(r.URL.Query().Get("limit"))
-	if err != nil || limit <= 0 {
-		limit = 50 // Default limit
-	}
-	if limit > 100 {
-		limit = 100 // Max limit
-	}
-
-	offset, err := strconv.Atoi(r.URL.Query().Get("offset"))
-	if err != nil || offset < 0 {
-		offset = 0
-	}
-
-	// Execute list query
-	workflows, total, err := h.workflowRepo.List(r.Context(), filters, limit, offset)
-	if err != nil {
-		h.logger.Error(err, "Failed to list workflows",
-			"filters", filters,
-			"limit", limit,
-			"offset", offset,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to list workflows", h.logger)
-		return
-	}
-
-	// Log success
-	h.logger.Info("Workflows listed",
-		"count", len(workflows),
-		"filters", filters,
-		"limit", limit,
-		"offset", offset,
-	)
-
-	// Convert to pointer slice for response
-	workflowPtrs := make([]*models.RemediationWorkflow, len(workflows))
-	for i := range workflows {
-		workflowPtrs[i] = &workflows[i]
-	}
-
-	// Return results
-	response := models.WorkflowListResponse{
-		Workflows: workflowPtrs,
-		Limit:     limit,
-		Offset:    offset,
-		Total:     total,
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(response); err != nil {
-		h.logger.Error(err, "Failed to encode workflow list response")
-	}
-}
-
-// HandleGetWorkflowByID handles GET /api/v1/workflows/{workflowID}
-// BR-STORAGE-039: Workflow Catalog Retrieval API
-// DD-WORKFLOW-002 v3.0: UUID primary key for workflow retrieval
-// DD-WORKFLOW-016, DD-HAPI-017: Security gate via optional context filters (Step 3)
-//
-// Returns complete workflow object including:
-// - spec.schema_image: OCI container image reference (nil for inline registrations)
-// - spec.parameters[]: Parameter schema (for LLM parameter validation)
-// - detected_labels: Signal type, severity labels (for workflow filtering)
-//
-// Security Gate (when context filters are present):
-// - Returns 404 if workflow exists but doesn't match context (DD-WORKFLOW-016)
-// - Intentionally doesn't distinguish "not found" from "filtered out" to prevent info leakage
-// - Emits workflow.catalog.workflow_retrieved audit event
-//
-// Cross-Service Integration:
-// - KA: Uses for get_workflow tool (Step 3 of discovery protocol)
-// - AIAnalysis: May use for defense-in-depth validation
-func (h *Handler) HandleGetWorkflowByID(w http.ResponseWriter, r *http.Request) {
-	// Get workflow ID from URL path
-	workflowID := chi.URLParam(r, "workflowID")
-	if workflowID == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"workflow_id is required", h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-016: Parse optional context filters for security gate
-	filters, err := ParseDiscoveryFilters(r)
-	if err != nil {
-		h.logger.Error(err, "Invalid discovery filter parameters")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"invalid discovery filter parameters", h.logger)
-		return
-	}
-
-	var wf *models.RemediationWorkflow
-
-	// DD-WORKFLOW-016: Use context-filtered query when filters are present (security gate)
-	// GAP-WF-6: Measure query duration for audit payload (DD-WORKFLOW-014 v3.0)
-	startGet := time.Now()
-	if filters.HasContextFilters() {
-		wf, err = h.workflowRepo.GetWorkflowWithContextFilters(r.Context(), workflowID, filters)
-	} else {
-		wf, err = h.workflowRepo.GetByID(r.Context(), workflowID)
-	}
-	durationMs := time.Since(startGet).Milliseconds()
-
-	if err != nil {
-		// Check if workflow not found
-		if err.Error() == fmt.Sprintf("workflow not found: %s", workflowID) {
-			response.WriteRFC7807Error(w, http.StatusNotFound, "workflow-not-found", "Not Found",
-				fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-			return
-		}
-
-		h.logger.Error(err, "Failed to get workflow",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to get workflow", h.logger)
-		return
-	}
-
-	// Check for nil workflow (not found or security gate filtered out)
-	if wf == nil {
-		h.logger.Info("Workflow not found or filtered by security gate",
-			"workflow_id", workflowID,
-			"has_context_filters", filters.HasContextFilters(),
-		)
-		// DD-WORKFLOW-016: Return same 404 for "not found" and "filtered out" (prevent info leakage)
-		response.WriteRFC7807Error(w, http.StatusNotFound, "workflow-not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-
-	// BR-AUDIT-023: Emit discovery audit events when context filters are present.
-	// DD-WORKFLOW-014 v3.0: Context filters indicate KA is validating its selection,
-	// so we emit both workflow_retrieved and selection_validated.
-	if filters.HasContextFilters() && h.auditStore != nil {
-		retrievedEvent, err := dsaudit.NewWorkflowRetrievedAuditEvent(workflowID, filters, durationMs)
-		if err != nil {
-			h.logger.Error(err, "Failed to create workflow_retrieved audit event", "workflow_id", workflowID)
-		}
-		validatedEvent, err := dsaudit.NewSelectionValidatedAuditEvent(workflowID, filters, true, durationMs)
-		if err != nil {
-			h.logger.Error(err, "Failed to create selection_validated audit event", "workflow_id", workflowID)
-		}
-
-		var events []*api.AuditEventRequest
-		if retrievedEvent != nil {
-			events = append(events, retrievedEvent)
-		}
-		if validatedEvent != nil {
-			events = append(events, validatedEvent)
-		}
-		if len(events) > 0 {
-			h.emitAuditEventsAsync(events, "workflow_id", workflowID)
-		}
-	}
-
-	// Log success
-	h.logger.Info("Workflow retrieved",
-		"workflow_id", workflowID,
-		"workflow_name", wf.WorkflowName,
-		"version", wf.Version,
-		"has_context_filters", filters.HasContextFilters(),
-	)
-
-	// Return workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(wf); err != nil {
-		h.logger.Error(err, "Failed to encode workflow response")
-	}
-}
-
-// HandleUpdateWorkflow handles PATCH /api/v1/workflows/{workflowID}
-// DD-WORKFLOW-012: Update ONLY mutable fields (status, metrics)
-// Immutable fields (description, content, labels) require creating a new version
-func (h *Handler) HandleUpdateWorkflow(w http.ResponseWriter, r *http.Request) {
-	// Get workflow ID from URL path
-	workflowID := chi.URLParam(r, "workflowID")
-	if workflowID == "" {
-		h.logger.Error(fmt.Errorf("workflow_id is required"), "Missing workflow_id in request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"workflow_id is required", h.logger)
-		return
-	}
-
-	// Parse request body
-	var updateReq models.WorkflowUpdateRequest
-	if err := json.NewDecoder(r.Body).Decode(&updateReq); err != nil {
-		if dsmiddleware.IsMaxBytesError(err) {
-			dsmiddleware.WriteMaxBytesExceeded(w, h.logger)
-			return
-		}
-		h.logger.Error(err, "Failed to decode workflow update request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"request body is not valid JSON", h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-012: Validate that ONLY mutable fields are being updated
-	if updateReq.Description != nil || updateReq.Content != nil || updateReq.Labels != nil {
-		h.logger.Error(fmt.Errorf("immutable fields cannot be updated"), "Attempted to update immutable fields",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "immutable-field-violation", "Bad Request",
-			"Cannot update immutable fields (description, content, labels). Create a new version instead. See DD-WORKFLOW-012.", h.logger)
-		return
-	}
-
-	// Get existing workflow
-	workflow, err := h.workflowRepo.GetByID(r.Context(), workflowID)
-	if err != nil {
-		h.logger.Error(err, "Failed to get workflow for update",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-
-	// Apply mutable field updates
-	if updateReq.Status != nil {
-		if workflowCatalogStatusTransitionForbidden(workflow.Status, *updateReq.Status) {
-			response.WriteRFC7807Error(w, http.StatusConflict, "workflow-status-conflict", "Conflict",
-				fmt.Sprintf("invalid workflow status transition from %s to %s", workflow.Status, *updateReq.Status), h.logger)
-			return
-		}
-		workflow.Status = *updateReq.Status
-		if *updateReq.Status == "Disabled" {
-			now := time.Now()
-			workflow.DisabledAt = &now
-			workflow.DisabledBy = updateReq.DisabledBy
-			workflow.DisabledReason = updateReq.DisabledReason
-		}
-	}
-
-	// Update the workflow
-	if err := h.workflowRepo.UpdateStatus(r.Context(), workflow.WorkflowID, workflow.Version, workflow.Status, getStringValue(workflow.DisabledReason), getStringValue(workflow.DisabledBy)); err != nil {
-		h.logger.Error(err, "Failed to update workflow",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to update workflow", h.logger)
-		return
-	}
-
-	// BR-STORAGE-183: Audit workflow update (business logic operation)
-	// DD-AUDIT-002 V2.0.1: Workflow state changes are business logic
-	// Copy values before goroutine to avoid data race (Issue #674 Bug 11)
-	if h.auditStore != nil {
-		wfStatus := workflow.Status
-		wfID := workflow.WorkflowID
-		wfDisabledBy := getStringValue(workflow.DisabledBy)
-		wfDisabledReason := getStringValue(workflow.DisabledReason)
-		isDisabling := updateReq.Status != nil && *updateReq.Status == "Disabled"
-		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			updatedFields := api.WorkflowCatalogUpdatedFields{}
-			updatedFields.Status.SetTo(wfStatus)
-
-			if isDisabling {
-				if wfDisabledBy != "" {
-					updatedFields.DisabledBy.SetTo(wfDisabledBy)
-				}
-				if wfDisabledReason != "" {
-					updatedFields.DisabledReason.SetTo(wfDisabledReason)
-				}
-			}
-
-			auditEvent, err := dsaudit.NewWorkflowUpdatedAuditEvent(wfID, updatedFields)
-			if err != nil {
-				h.logger.Error(err, "Failed to create workflow update audit event",
-					"workflow_id", workflowID,
-				)
-				return
-			}
-
-			if err := h.auditStore.StoreAudit(ctx, auditEvent); err != nil {
-				h.logger.Error(err, "Failed to audit workflow update",
-					"workflow_id", workflowID,
-				)
-			}
-		}()
-	}
-
-	// Log success
-	h.logger.Info("Workflow updated",
-		"workflow_id", workflowID,
-		"status", workflow.Status,
-	)
-
-	// Return updated workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(workflow); err != nil {
-		h.logger.Error(err, "Failed to encode workflow response")
-	}
-}
-
-// getStringValue safely dereferences a string pointer
-func getStringValue(s *string) string {
-	if s == nil {
-		return ""
-	}
-	return *s
-}
-
-// emitAuditEventsAsync stores one or more audit events asynchronously in a background goroutine.
-// This is the standard pattern for non-blocking audit event emission (BR-AUDIT-024).
-// Events are stored sequentially within a single goroutine to share one context/timeout.
-// If event creation fails for any event, remaining events are still attempted.
-func (h *Handler) emitAuditEventsAsync(events []*api.AuditEventRequest, kvs ...interface{}) {
-	go func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		for _, event := range events {
-			if err := h.auditStore.StoreAudit(ctx, event); err != nil {
-				args := append([]interface{}{"Failed to store audit event", "event_type", event.EventType}, kvs...)
-				h.logger.Error(err, args[0].(string), args[1:]...)
-			}
-		}
-	}()
-}
-
-// workflowCatalogStatusTransitionForbidden returns true when a catalog status PATCH should be rejected.
-// Valid: Active→Disabled, Active→Deprecated, Active→Superseded, Disabled→Active.
-// Terminal states (DD-WORKFLOW-017): Superseded and Deprecated cannot transition to any other status.
-// Same-status transitions are no-ops and always allowed.
-func workflowCatalogStatusTransitionForbidden(fromStatus, toStatus string) bool {
-	if fromStatus == toStatus {
-		return false
-	}
-	return fromStatus == "Superseded" || fromStatus == "Deprecated"
-}
-
-// getWorkflowLifecycleRepo returns the workflow lifecycle repository for enable/disable/deprecate.
-// Uses workflowLifecycleRepo when set (tests), otherwise workflowRepo.
-func (h *Handler) getWorkflowLifecycleRepo() WorkflowLifecycleRepository {
-	if h.workflowLifecycleRepo != nil {
-		return h.workflowLifecycleRepo
-	}
-	return h.workflowRepo
-}
-
-// HandleEnableWorkflow handles PATCH /api/v1/workflows/{workflowID}/enable
-// DD-WORKFLOW-017 Phase 4.4 (GAP-WF-1): Convenience endpoint for re-enabling workflows
-func (h *Handler) HandleEnableWorkflow(w http.ResponseWriter, r *http.Request) {
-	// Get workflow ID from URL path
-	workflowID := chi.URLParam(r, "workflowID")
-	if workflowID == "" {
-		h.logger.Error(fmt.Errorf("workflow_id is required"), "Missing workflow_id in request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"workflow_id is required", h.logger)
-		return
-	}
-
-	// Parse request body
-	var req models.WorkflowDisableRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		if dsmiddleware.IsMaxBytesError(err) {
-			dsmiddleware.WriteMaxBytesExceeded(w, h.logger)
-			return
-		}
-		h.logger.Error(err, "Failed to decode workflow enable request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"request body is not valid JSON", h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-017 Phase 4.4 (GAP-WF-5): reason is mandatory for lifecycle operations
-	if req.Reason == nil || strings.TrimSpace(*req.Reason) == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "missing-reason", "Missing Required Field",
-			"reason is required for lifecycle operations", h.logger)
-		return
-	}
-
-	// Get existing workflow
-	repo := h.getWorkflowLifecycleRepo()
-	if repo == nil {
-		h.logger.Error(fmt.Errorf("workflow repository not configured"), "Handler misconfiguration")
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Workflow repository not configured", h.logger)
-		return
-	}
-
-	workflow, err := repo.GetByID(r.Context(), workflowID)
-	if err != nil {
-		h.logger.Error(err, "Failed to get workflow for enable",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-	if workflow == nil {
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-
-	if workflowCatalogStatusTransitionForbidden(workflow.Status, "Active") {
-		response.WriteRFC7807Error(w, http.StatusConflict, "workflow-status-conflict", "Conflict",
-			fmt.Sprintf("invalid workflow status transition from %s to Active", workflow.Status), h.logger)
-		return
-	}
-
-	// Update the workflow status to active
-	reason := ""
-	if req.Reason != nil {
-		reason = *req.Reason
-	}
-	updatedBy := ""
-	if req.UpdatedBy != nil {
-		updatedBy = *req.UpdatedBy
-	}
-
-	if err := repo.UpdateStatus(r.Context(), workflow.WorkflowID, workflow.Version, "Active", reason, updatedBy); err != nil {
-		h.logger.Error(err, "Failed to enable workflow",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to enable workflow", h.logger)
-		return
-	}
-
-	// Update workflow object for response
-	workflow.Status = "Active"
-	workflow.DisabledAt = nil
-	workflow.DisabledBy = nil
-	workflow.DisabledReason = nil
-
-	// BR-STORAGE-183: Audit workflow enable (business logic operation)
-	if h.auditStore != nil {
-		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			updatedFields := api.WorkflowCatalogUpdatedFields{}
-			updatedFields.Status.SetTo("Active")
-
-			auditEvent, err := dsaudit.NewWorkflowUpdatedAuditEvent(workflow.WorkflowID, updatedFields)
-			if err != nil {
-				h.logger.Error(err, "Failed to create workflow enable audit event",
-					"workflow_id", workflowID,
-				)
-				return
-			}
-
-			if err := h.auditStore.StoreAudit(ctx, auditEvent); err != nil {
-				h.logger.Error(err, "Failed to audit workflow enable",
-					"workflow_id", workflowID,
-				)
-			}
-		}()
-	}
-
-	// Log success
-	h.logger.Info("Workflow enabled",
-		"workflow_id", workflowID,
-		"reason", reason,
-		"updated_by", updatedBy,
-	)
-
-	// Return updated workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(workflow); err != nil {
-		h.logger.Error(err, "Failed to encode workflow response")
-	}
-}
-
-// HandleDeprecateWorkflow handles PATCH /api/v1/workflows/{workflowID}/deprecate
-// DD-WORKFLOW-017 Phase 4.4 (GAP-WF-1): Convenience endpoint for deprecating workflows
-func (h *Handler) HandleDeprecateWorkflow(w http.ResponseWriter, r *http.Request) {
-	// Get workflow ID from URL path
-	workflowID := chi.URLParam(r, "workflowID")
-	if workflowID == "" {
-		h.logger.Error(fmt.Errorf("workflow_id is required"), "Missing workflow_id in request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"workflow_id is required", h.logger)
-		return
-	}
-
-	// Parse request body
-	var req models.WorkflowDisableRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		if dsmiddleware.IsMaxBytesError(err) {
-			dsmiddleware.WriteMaxBytesExceeded(w, h.logger)
-			return
-		}
-		h.logger.Error(err, "Failed to decode workflow deprecate request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"request body is not valid JSON", h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-017 Phase 4.4 (GAP-WF-5): reason is mandatory for lifecycle operations
-	if req.Reason == nil || strings.TrimSpace(*req.Reason) == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "missing-reason", "Missing Required Field",
-			"reason is required for lifecycle operations", h.logger)
-		return
-	}
-
-	// Get existing workflow
-	repo := h.getWorkflowLifecycleRepo()
-	if repo == nil {
-		h.logger.Error(fmt.Errorf("workflow repository not configured"), "Handler misconfiguration")
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Workflow repository not configured", h.logger)
-		return
-	}
-
-	workflow, err := repo.GetByID(r.Context(), workflowID)
-	if err != nil {
-		h.logger.Error(err, "Failed to get workflow for deprecate",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-	if workflow == nil {
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-017: Superseded and Deprecated are terminal states
-	if workflowCatalogStatusTransitionForbidden(workflow.Status, "Deprecated") {
-		response.WriteRFC7807Error(w, http.StatusConflict, "workflow-status-conflict", "Status Transition Forbidden",
-			fmt.Sprintf("Cannot transition from %s to Deprecated", workflow.Status), h.logger)
-		return
-	}
-
-	// Update the workflow status to deprecated
-	reason := ""
-	if req.Reason != nil {
-		reason = *req.Reason
-	}
-	updatedBy := ""
-	if req.UpdatedBy != nil {
-		updatedBy = *req.UpdatedBy
-	}
-
-	if err := repo.UpdateStatus(r.Context(), workflow.WorkflowID, workflow.Version, "Deprecated", reason, updatedBy); err != nil {
-		h.logger.Error(err, "Failed to deprecate workflow",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to deprecate workflow", h.logger)
-		return
-	}
-
-	// Update workflow object for response
-	workflow.Status = "Deprecated"
-
-	// BR-STORAGE-183: Audit workflow deprecate (business logic operation)
-	if h.auditStore != nil {
-		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			updatedFields := api.WorkflowCatalogUpdatedFields{}
-			updatedFields.Status.SetTo("Deprecated")
-
-			auditEvent, err := dsaudit.NewWorkflowUpdatedAuditEvent(workflow.WorkflowID, updatedFields)
-			if err != nil {
-				h.logger.Error(err, "Failed to create workflow deprecate audit event",
-					"workflow_id", workflowID,
-				)
-				return
-			}
-
-			if err := h.auditStore.StoreAudit(ctx, auditEvent); err != nil {
-				h.logger.Error(err, "Failed to audit workflow deprecate",
-					"workflow_id", workflowID,
-				)
-			}
-		}()
-	}
-
-	// Log success
-	h.logger.Info("Workflow deprecated",
-		"workflow_id", workflowID,
-		"reason", reason,
-		"updated_by", updatedBy,
-	)
-
-	// Return updated workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(workflow); err != nil {
-		h.logger.Error(err, "Failed to encode workflow response")
-	}
-}
-
-// HandleDisableWorkflow handles PATCH /api/v1/workflows/{workflowID}/disable
-// DD-WORKFLOW-012: Convenience endpoint for disabling workflows (soft delete)
-func (h *Handler) HandleDisableWorkflow(w http.ResponseWriter, r *http.Request) {
-	// Get workflow ID from URL path
-	workflowID := chi.URLParam(r, "workflowID")
-	if workflowID == "" {
-		h.logger.Error(fmt.Errorf("workflow_id is required"), "Missing workflow_id in request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"workflow_id is required", h.logger)
-		return
-	}
-
-	// Parse request body
-	var disableReq models.WorkflowDisableRequest
-	if err := json.NewDecoder(r.Body).Decode(&disableReq); err != nil {
-		if dsmiddleware.IsMaxBytesError(err) {
-			dsmiddleware.WriteMaxBytesExceeded(w, h.logger)
-			return
-		}
-		h.logger.Error(err, "Failed to decode workflow disable request")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"request body is not valid JSON", h.logger)
-		return
-	}
-
-	// DD-WORKFLOW-017 Phase 4.4 (GAP-WF-5): reason is mandatory for lifecycle operations
-	if disableReq.Reason == nil || strings.TrimSpace(*disableReq.Reason) == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "missing-reason", "Missing Required Field",
-			"reason is required for lifecycle operations", h.logger)
-		return
-	}
-
-	// Get existing workflow
-	repo := h.getWorkflowLifecycleRepo()
-	if repo == nil {
-		h.logger.Error(fmt.Errorf("workflow repository not configured"), "Handler misconfiguration")
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Workflow repository not configured", h.logger)
-		return
-	}
-
-	workflow, err := repo.GetByID(r.Context(), workflowID)
-	if err != nil {
-		h.logger.Error(err, "Failed to get workflow for disable",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-	if workflow == nil {
-		response.WriteRFC7807Error(w, http.StatusNotFound, "not-found", "Not Found",
-			fmt.Sprintf("Workflow not found: %s", workflowID), h.logger)
-		return
-	}
-
-	if workflowCatalogStatusTransitionForbidden(workflow.Status, "Disabled") {
-		response.WriteRFC7807Error(w, http.StatusConflict, "workflow-status-conflict", "Conflict",
-			fmt.Sprintf("invalid workflow status transition from %s to Disabled", workflow.Status), h.logger)
-		return
-	}
-
-	// Update the workflow status to disabled
-	reason := ""
-	if disableReq.Reason != nil {
-		reason = *disableReq.Reason
-	}
-	updatedBy := ""
-	if disableReq.UpdatedBy != nil {
-		updatedBy = *disableReq.UpdatedBy
-	}
-
-	if err := repo.UpdateStatus(r.Context(), workflow.WorkflowID, workflow.Version, "Disabled", reason, updatedBy); err != nil {
-		h.logger.Error(err, "Failed to disable workflow",
-			"workflow_id", workflowID,
-		)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to disable workflow", h.logger)
-		return
-	}
-
-	// Update workflow object for response
-	workflow.Status = "Disabled"
-	now := time.Now()
-	workflow.DisabledAt = &now
-	workflow.DisabledBy = disableReq.UpdatedBy
-	workflow.DisabledReason = disableReq.Reason
-
-	// BR-STORAGE-183: Audit workflow disable (business logic operation)
-	// DD-AUDIT-002 V2.0.1: Workflow disable is a status update (captured via workflow.updated)
-	if h.auditStore != nil {
-		go func() {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			// Build structured updated fields (OGEN-MIGRATION)
-			updatedFields := api.WorkflowCatalogUpdatedFields{}
-			updatedFields.Status.SetTo("Disabled")
-			updatedFields.DisabledBy.SetTo(updatedBy)
-			updatedFields.DisabledReason.SetTo(reason)
-
-			auditEvent, err := dsaudit.NewWorkflowUpdatedAuditEvent(workflow.WorkflowID, updatedFields)
-			if err != nil {
-				h.logger.Error(err, "Failed to create workflow disable audit event",
-					"workflow_id", workflowID,
-				)
-				return
-			}
-
-			if err := h.auditStore.StoreAudit(ctx, auditEvent); err != nil {
-				h.logger.Error(err, "Failed to audit workflow disable",
-					"workflow_id", workflowID,
-				)
-			}
-		}()
-	}
-
-	// Log success
-	h.logger.Info("Workflow disabled",
-		"workflow_id", workflowID,
-		"reason", reason,
-		"disabled_by", updatedBy,
-	)
-
-	// Return updated workflow
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(workflow); err != nil {
-		h.logger.Error(err, "Failed to encode workflow response")
-	}
-}
-
-// ========================================
-// THREE-STEP WORKFLOW DISCOVERY HANDLERS
-// ========================================
-// Authority: DD-WORKFLOW-016 (Action-Type Workflow Catalog Indexing)
-// Authority: DD-HAPI-017 (Three-Step Workflow Discovery Integration)
-// Business Requirement: BR-HAPI-017-001 (Three-Step Tool Implementation)
-//
-// Step 1: GET /api/v1/workflows/actions - List available action types
-// Step 2: GET /api/v1/workflows/actions/{action_type} - List workflows for action type
-// Step 3: GET /api/v1/workflows/{workflowID} (modified) - Get workflow with security gate
-
-// HandleListAvailableActions handles GET /api/v1/workflows/actions
-// Step 1: Returns action types from taxonomy that have active workflows
-// matching the provided signal context filters.
-// Emits workflow.catalog.actions_listed audit event (DD-WORKFLOW-014 v3.0)
-func (h *Handler) HandleListAvailableActions(w http.ResponseWriter, r *http.Request) {
-	// Parse discovery filters from query parameters
-	filters, err := ParseDiscoveryFilters(r)
-	if err != nil {
-		h.logger.Error(err, "Invalid discovery filter parameters")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"invalid discovery filter parameters", h.logger)
-		return
-	}
-
-	// Parse pagination
-	offset, limit := ParsePagination(r)
-
-	// Execute query (GAP-WF-6: measure duration for audit payload)
-	startList := time.Now()
-	entries, totalCount, err := h.workflowRepo.ListActions(r.Context(), filters, offset, limit)
-	durationMs := time.Since(startList).Milliseconds()
-	if err != nil {
-		h.logger.Error(err, "Failed to list available actions",
-			"filters", filters)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to list available actions", h.logger)
-		return
-	}
-
-	// Build response
-	resp := models.ActionTypeListResponse{
-		ActionTypes: entries,
-		Pagination: models.PaginationMetadata{
-			TotalCount: totalCount,
-			Offset:     offset,
-			Limit:      limit,
-			HasMore:    offset+limit < totalCount,
-		},
-	}
-
-	// BR-AUDIT-023: Emit workflow.catalog.actions_listed audit event
-	if h.auditStore != nil {
-		if event, err := dsaudit.NewActionsListedAuditEvent(filters, totalCount, durationMs); err != nil {
-			h.logger.Error(err, "Failed to create actions_listed audit event")
-		} else {
-			h.emitAuditEventsAsync([]*api.AuditEventRequest{event})
-		}
-	}
-
-	h.logger.Info("Available actions listed",
-		"total_count", totalCount,
-		"returned_count", len(entries),
-		"offset", offset,
-		"limit", limit,
-	)
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(resp); err != nil {
-		h.logger.Error(err, "Failed to encode actions list response")
-	}
-}
-
-// HandleListWorkflowsByActionType handles GET /api/v1/workflows/actions/{action_type}
-// Step 2: Returns active workflows matching the specified action type and context filters.
-// Emits workflow.catalog.workflows_listed audit event (DD-WORKFLOW-014 v3.0)
-func (h *Handler) HandleListWorkflowsByActionType(w http.ResponseWriter, r *http.Request) {
-	// Get action_type from URL path
-	actionType := chi.URLParam(r, "action_type")
-	if actionType == "" {
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"action_type path parameter is required", h.logger)
-		return
-	}
-
-	// Parse discovery filters from query parameters
-	filters, err := ParseDiscoveryFilters(r)
-	if err != nil {
-		h.logger.Error(err, "Invalid discovery filter parameters")
-		response.WriteRFC7807Error(w, http.StatusBadRequest, "bad-request", "Bad Request",
-			"invalid discovery filter parameters", h.logger)
-		return
-	}
-
-	// Parse pagination
-	offset, limit := ParsePagination(r)
-
-	// Execute query (GAP-WF-6: measure duration for audit payload)
-	startList := time.Now()
-	workflows, totalCount, err := h.workflowRepo.ListWorkflowsByActionType(r.Context(), actionType, filters, offset, limit)
-	durationMs := time.Since(startList).Milliseconds()
-	if err != nil {
-		h.logger.Error(err, "Failed to list workflows by action type",
-			"action_type", actionType)
-		response.WriteRFC7807Error(w, http.StatusInternalServerError, "internal-error", "Internal Server Error",
-			"Failed to list workflows by action type", h.logger)
-		return
-	}
-
-	// Convert to discovery entries
-	// DD-HAPI-017 v1.1: ActualSuccessRate and TotalExecutions excluded from
-	// LLM-facing response — global aggregates are misleading for per-incident selection.
-	discoveryEntries := make([]models.WorkflowDiscoveryEntry, 0, len(workflows))
-	for _, wf := range workflows {
-		entry := models.WorkflowDiscoveryEntry{
-			WorkflowID:      wf.WorkflowID,
-			WorkflowName:    wf.WorkflowName,
-			Name:            wf.Name,
-			Description:     wf.Description,
-			Version:         wf.Version,
-			ExecutionEngine: string(wf.ExecutionEngine),
-		}
-		if wf.SchemaImage != nil {
-			entry.SchemaImage = *wf.SchemaImage
-		}
-		if wf.ExecutionBundle != nil {
-			entry.ExecutionBundle = *wf.ExecutionBundle
-		}
-		if wf.ServiceAccountName != nil {
-			entry.ServiceAccountName = *wf.ServiceAccountName
-		}
-		discoveryEntries = append(discoveryEntries, entry)
-	}
-
-	// Build response
-	resp := models.WorkflowDiscoveryResponse{
-		ActionType: actionType,
-		Workflows:  discoveryEntries,
-		Pagination: models.PaginationMetadata{
-			TotalCount: totalCount,
-			Offset:     offset,
-			Limit:      limit,
-			HasMore:    offset+limit < totalCount,
-		},
-	}
-
-	// BR-AUDIT-023: Emit workflow.catalog.workflows_listed audit event
-	if h.auditStore != nil {
-		if event, err := dsaudit.NewWorkflowsListedAuditEvent(actionType, filters, totalCount, durationMs); err != nil {
-			h.logger.Error(err, "Failed to create workflows_listed audit event", "action_type", actionType)
-		} else {
-			h.emitAuditEventsAsync([]*api.AuditEventRequest{event}, "action_type", actionType)
-		}
-	}
-
-	h.logger.Info("Workflows listed by action type",
-		"action_type", actionType,
-		"total_count", totalCount,
-		"returned_count", len(discoveryEntries),
-		"offset", offset,
-		"limit", limit,
-	)
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(resp); err != nil {
-		h.logger.Error(err, "Failed to encode workflows list response")
-	}
-}
-
-// ParseDiscoveryFilters extracts WorkflowDiscoveryFilters from query parameters.
-// Used by all three discovery endpoints.
-// Exported for unit testing (UT-DS-017-001-003).
-func ParseDiscoveryFilters(r *http.Request) (*models.WorkflowDiscoveryFilters, error) {
-	filters := &models.WorkflowDiscoveryFilters{
-		Severity:      r.URL.Query().Get("severity"),
-		Component:     r.URL.Query().Get("component"),
-		Environment:   r.URL.Query().Get("environment"),
-		Priority:      r.URL.Query().Get("priority"),
-		RemediationID: r.URL.Query().Get("remediation_id"),
-	}
-
-	// Parse optional JSON-encoded custom_labels
-	if customLabelsStr := r.URL.Query().Get("custom_labels"); customLabelsStr != "" {
-		var customLabels map[string][]string
-		if err := json.Unmarshal([]byte(customLabelsStr), &customLabels); err != nil {
-			return nil, fmt.Errorf("invalid custom_labels JSON: %w", err)
-		}
-		filters.CustomLabels = customLabels
-	}
-
-	// Parse optional JSON-encoded detected_labels
-	if detectedLabelsStr := r.URL.Query().Get("detected_labels"); detectedLabelsStr != "" {
-		var detectedLabels models.DetectedLabels
-		if err := json.Unmarshal([]byte(detectedLabelsStr), &detectedLabels); err != nil {
-			return nil, fmt.Errorf("invalid detected_labels JSON: %w", err)
-		}
-		filters.DetectedLabels = &detectedLabels
-	}
-
-	return filters, nil
-}
-
-// ParsePagination extracts offset and limit from query parameters.
-// Defaults: offset=0, limit=10 (DD-WORKFLOW-016 default page size)
-// Exported for unit testing (UT-DS-017-001-002).
-func ParsePagination(r *http.Request) (int, int) {
-	offset, err := strconv.Atoi(r.URL.Query().Get("offset"))
-	if err != nil || offset < 0 {
-		offset = 0
-	}
-
-	limit, err := strconv.Atoi(r.URL.Query().Get("limit"))
-	if err != nil || limit <= 0 {
-		limit = models.DefaultPaginationLimit
-	}
-	if limit > models.MaxPaginationLimit {
-		limit = models.MaxPaginationLimit
-	}
-
-	return offset, limit
-}
-
-// computeContentHash computes a SHA-256 hash of the workflow content.
-// DD-WORKFLOW-017: Content hash is derived from the raw YAML content (inline or OCI).
-func computeContentHash(content string) string {
-	hash := sha256.Sum256([]byte(content))
-	return hex.EncodeToString(hash[:])
-}
